% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_in_str.R
\name{filter_in_str}
\alias{filter_in_str}
\title{Filter a column of a data frame by a string vector
It will return a vector of boolean}
\usage{
filter_in_str(vec, ind_v)
}
\arguments{
\item{vec}{Column of the dataframe you want to filter.}

\item{y}{A string vector.}
}
\value{
Boolean that represents if the column include the vector or not.
}
\description{
Filter a column of a data frame by a string vector
It will return a vector of boolean
}
\examples{
filter_in_str(iris$Species,c("est","sico"))
}
