% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pratcTidy.R
\docType{package}
\name{pratcTidy-package}
\alias{pratcTidy}
\alias{pratcTidy-package}
\title{pratcTidy: Tidy data in practical way (Title Case)}
\description{
To learn more about financialAdventure, start with the vignettes:
browseVignettes(package = "pratcTidy")
}
\seealso{
Useful links:
\itemize{
  \item \url{https://medicihouse07.github.io/}
}

}
\author{
\strong{Maintainer}: Medici House \email{medicihouse07@gmail.com} (\href{https://orcid.org/0000-0002-8572-3755}{ORCID})

}
\keyword{internal}
